rearrange classes pt 2 (test)
